feat #3970 Move methods to JournalManager
