+++ src/journal/models.py
@@ -35,6 +35,7 @@ from django.utils import timezone, translation
 from django.utils.functional import cached_property
 from django.utils.html import strip_tags
 from django.utils.translation import gettext, gettext_lazy as _
+from modeltranslation.utils import build_localized_fieldname
 
 from core import (
     files,
@@ -98,6 +99,97 @@ def issue_large_image_path(instance, filename):
 
     path = "issues/{0}".format(instance.pk)
     return os.path.join(path, filename)
+
+
+class JournalManager(models.Manager):
+    def _apply_ordering_az(self, journals):
+        """
+        Order a queryset of journals A-Z on English-language journal name.
+        Note that this does not support multilingual journal names:
+        more work is needed on django-modeltranslation to
+        support Django subqueries.
+        :param journals: Queryset of Journal objects
+        """
+        localized_column = build_localized_fieldname(
+            "value",
+            settings.LANGUAGE_CODE,  # Assumed to be 'en' in default config
+        )
+        name = core_models.SettingValue.objects.filter(
+            journal=models.OuterRef("pk"),
+            setting__name="journal_name",
+        )
+        journals = journals.annotate(
+            journal_name=models.Subquery(
+                name.values_list(localized_column, flat=True)[:1],
+                output_field=models.CharField(),
+            )
+        )
+        return journals.order_by("journal_name")
+
+    def _apply_ordering(self, journals):
+        press = press_models.Press.objects.all().first()
+        if press.order_journals_az:
+            return self._apply_ordering_az(journals)
+        else:
+            # Journals will already have been ordered according to Meta.ordering
+            return journals
+
+    @property
+    def public_journals(self):
+        """
+        Get all journals that are not hidden from the press
+        or designated as conferences.
+        Do not apply ordering yet,
+        since the caller may filter the queryset.
+        """
+        return self.get_queryset().filter(
+            hide_from_press=False,
+            is_conference=False,
+        )
+
+    @property
+    def public_active_journals(self):
+        """
+        Get all journals that are visible to the press
+        and marked as 'Active' or 'Test' in the publishing status field.
+
+        Note: Test journals are included so that users can test the journal
+        list safely. A separate mechanism exists to hide them from the press
+        once the press enters normal operation:
+        Journal.hide_from_press.
+        """
+        return self._apply_ordering(
+            self.public_journals.filter(
+                status__in=[
+                    Journal.PublishingStatus.ACTIVE,
+                    Journal.PublishingStatus.TEST,
+                ]
+            )
+        )
+
+    @property
+    def public_archived_journals(self):
+        """
+        Get all journals that are visible to the press
+        and marked as 'Archived' in the publishing status field.
+        """
+        return self._apply_ordering(
+            self.public_journals.filter(
+                status=Journal.PublishingStatus.ARCHIVED,
+            )
+        )
+
+    @property
+    def public_coming_soon_journals(self):
+        """
+        Get all journals that are visible to the press
+        and marked as 'Coming soon' in the publishing status field.
+        """
+        return self._apply_ordering(
+            self.public_journals.filter(
+                status=Journal.PublishingStatus.COMING_SOON,
+            )
+        )
 
 
 class Journal(AbstractSiteModel):
@@ -313,6 +405,8 @@ class Journal(AbstractSiteModel):
     )
 
     disable_front_end = models.BooleanField(default=False)
+
+    objects = JournalManager()
 
     class Meta:
         ordering = ("sequence",)
